refactor: split daily memory runtime helpers
