+++ packages/memory-host-sdk/src/host/json-files.ts
@@ -1,0 +1,93 @@
+import { randomUUID } from "node:crypto";
+import fs from "node:fs/promises";
+import path from "node:path";
+
+function getErrorCode(err: unknown): string | undefined {
+  return err instanceof Error ? (err as NodeJS.ErrnoException).code : undefined;
+}
+
+async function replaceFileWithWindowsFallback(tempPath: string, filePath: string, mode: number) {
+  try {
+    await fs.rename(tempPath, filePath);
+    return;
+  } catch (err) {
+    const code = getErrorCode(err);
+    if (process.platform !== "win32" || (code !== "EPERM" && code !== "EEXIST")) {
+      throw err;
+    }
+  }
+
+  const existing = await fs.lstat(filePath).catch(() => null);
+  if (existing?.isSymbolicLink()) {
+    await fs.rm(filePath, { force: true });
+    await fs.rename(tempPath, filePath);
+    return;
+  }
+
+  await fs.copyFile(tempPath, filePath);
+  try {
+    await fs.chmod(filePath, mode);
+  } catch {}
+  await fs.rm(tempPath, { force: true }).catch(() => undefined);
+}
+
+export async function writeJsonAtomic(
+  filePath: string,
+  value: unknown,
+  options?: { mode?: number; trailingNewline?: boolean; ensureDirMode?: number },
+) {
+  const text = JSON.stringify(value, null, 2);
+  const mode = options?.mode ?? 0o600;
+  const payload = options?.trailingNewline && !text.endsWith("\n") ? `${text}\n` : text;
+  const mkdirOptions: { recursive: true; mode?: number } = { recursive: true };
+  if (typeof options?.ensureDirMode === "number") {
+    mkdirOptions.mode = options.ensureDirMode;
+  }
+
+  await fs.mkdir(path.dirname(filePath), mkdirOptions);
+  const parentDir = path.dirname(filePath);
+  const tmp = `${filePath}.${randomUUID()}.tmp`;
+  try {
+    const tmpHandle = await fs.open(tmp, "w", mode);
+    try {
+      await tmpHandle.writeFile(payload, { encoding: "utf8" });
+      await tmpHandle.sync();
+    } finally {
+      await tmpHandle.close().catch(() => undefined);
+    }
+    try {
+      await fs.chmod(tmp, mode);
+    } catch {}
+    await replaceFileWithWindowsFallback(tmp, filePath, mode);
+    try {
+      const dirHandle = await fs.open(parentDir, "r");
+      try {
+        await dirHandle.sync();
+      } finally {
+        await dirHandle.close().catch(() => undefined);
+      }
+    } catch {}
+    try {
+      await fs.chmod(filePath, mode);
+    } catch {}
+  } finally {
+    await fs.rm(tmp, { force: true }).catch(() => undefined);
+  }
+}
+
+export function createAsyncLock() {
+  let lock: Promise<void> = Promise.resolve();
+  return async function withLock<T>(fn: () => Promise<T>): Promise<T> {
+    const prev = lock;
+    let release: (() => void) | undefined;
+    lock = new Promise<void>((resolve) => {
+      release = resolve;
+    });
+    await prev;
+    try {
+      return await fn();
+    } finally {
+      release?.();
+    }
+  };
+}
